Moved function to separate file
